Update .gitignore to exclude net8.0 Debug directory
Updated the `.gitignore` file to ignore the `/src/FSMS/FSMS/bin/Debug/net8.0` directory. This change ensures that Git will not track or include any changes in this directory in commits. This is a standard practice for build output directories, as their contents are generated from the source code and do not need to be stored in the source code repository.