Update .gitignore with new path
Updated the `.gitignore` file to include the path `/src/FSMS/.vs/FSMS/v17/.suo`. This change ensures that Git will ignore this file or directory, preventing tracking of changes. This is typically done for files generated during the build process or specific to a developer's environment, which should not be included in the shared repository.