Added new Document object and updated DocumentGroups properties
Added a new Document object to the `Documents` array in `DocumentLayout.json` with properties providing information about the document's location and identity. Updated the `SelectedChildIndex` property in the `DocumentGroups` array from `-1` to `7`, indicating a change in the selected child in the document group. Also, added a new Document object to the `Children` array in the `DocumentGroups` array with properties providing information about the document's index, title, location, state, icon, opening time, and editor caption.